pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "promptracer"
version = "0.3.0"
description = "The open-source prompt engineering toolkit. Test, compare, and evaluate prompts across any LLM."
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
authors = [
    { name = "Cesar Segura", email = "cesarfsegurap09@gmail.com" },
]
keywords = ["llm", "prompt-engineering", "ai", "openai", "anthropic", "gemini", "ollama"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    "httpx>=0.27",
    "typer>=0.12",
    "rich>=13.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
openai = ["openai>=1.0"]
anthropic = ["anthropic>=0.39"]
google = ["google-genai>=1.0"]
all = ["openai>=1.0", "anthropic>=0.39", "google-genai>=1.0"]
dev = ["pytest>=8.0", "pytest-asyncio>=0.24", "ruff>=0.8"]

[project.scripts]
promptracer = "promptracer.cli:app"

[project.urls]
Homepage = "https://github.com/Unlucko/promptracer"
Repository = "https://github.com/Unlucko/promptracer"
Issues = "https://github.com/Unlucko/promptracer/issues"

[tool.hatch.build.targets.wheel]
packages = ["src/promptracer"]

[tool.ruff]
target-version = "py310"
line-length = 100

[tool.pytest.ini_options]
testpaths = ["tests"]
